# src/config/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Wallet-App",
        "description": "API Documentations for Simple Wallet App with Express.js.",
        "license": {
            "name": "myname",
            "url": "https://opensource.org/licenses/myname"
        }
    },
    "host": "localhost:3000",
    "basePath": "/",
    "tags": [
        {
            "name": "Customers",
            "description": "API for customers in the system"
        }
    ],
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
        "/customers": {
            "post": {
                "tags": [
                    "Customers"
                ],
                "description": "Create new customer in system",
                "summary": "Create new customer.",
                "parameters": [
                    {
                        "name": "customer",
                        "in": "body",
                        "description": "Customer that we want to create",
                        "schema": {
                            "$ref": "#/definitions/Customer"
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "201": {
                        "description": "Customer is created",
                        "schema": {
                            "$ref": "#/definitions/CustomerDTO"
                        }
                    }
                }
            },
            "put": {
                "summary": "Update customer with give ID",
                "tags": [
                    "Customers"
                ],
                "parameters": [
                    {
                        "name": "CustomerDTO",
                        "in": "body",
                        "description": "Customer with new values of properties",
                        "schema": {
                            "$ref": "#/definitions/CustomerDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Customer is updated",
                        "schema": {
                            "$ref": "#/definitions/CustomerDTO"
                        }
                    }
                }
            },
            "get": {
                "tags": [
                    "Customers"
                ],
                "summary": "Get list of customers.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Customers"
                        }
                    }
                }
            }
        },
        "/customers/{id}": {
            "parameters": [
                {
                    "name": "id",
                    "in": "path",
                    "required": true,
                    "description": "ID of customer that we want to find",
                    "type": "string"
                }
            ],
            "get": {
                "tags": [
                    "Customers"
                ],
                "summary": "Get customer with given ID",
                "responses": {
                    "200": {
                        "description": "Customer is found",
                        "schema": {
                            "$ref": "#/definitions/CustomerDTO"
                        }
                    }
                }
            },
            "delete": {
                "summary": "Delete customer with given ID",
                "tags": [
                    "Customers"
                ],
                "responses": {
                    "200": {
                        "description": "Customer is deleted",
                        "schema": {
                            "$ref": "#/definitions/CustomerDTO"
                        }
                    }
                }
            }
        },
        "/transactions": {
            "post": {
                "tags": [
                    "Transactions"
                ],
                "description": "Create new transaction in system",
                "summary": "Create new transaction.",
                "parameters": [
                    {
                        "name": "transaction",
                        "in": "body",
                        "description": "Transaction that we want to create",
                        "schema": {
                            "$ref": "#/definitions/Transaction"
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "201": {
                        "description": "Transaction is created",
                        "schema": {
                            "$ref": "#/definitions/Transaction"
                        }
                    }
                }
            },
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Get list of transactions.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Transactions"
                        }
                    }
                }
            }
        },
        "/transactions/history": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Get list of customer transactions.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Transactions"
                        }
                    }
                }
            }
        },
        "/transactions/{id}": {
            "parameters": [
                {
                    "name": "id",
                    "in": "path",
                    "required": true,
                    "description": "ID of transaction that we want to find",
                    "type": "string"
                }
            ],
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Get transaction with given ID",
                "responses": {
                    "200": {
                        "description": "Transaction is found",
                        "schema": {
                            "$ref": "#/definitions/Transaction"
                        }
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "description": "Login to system.",
                "summary": "Login to system.",
                "parameters": [
                    {
                        "name": "data",
                        "in": "body",
                        "description": "Login credentials.",
                        "schema": {
                            "$ref": "#/definitions/Login"
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Login success",
                        "schema": {
                            "$ref": "#/definitions/Customer"
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout from the system",
                "responses": {
                    "200": {
                        "description": "Logout success"
                    }
                }
            }
        }
    },
    "definitions": {
        "Customer": {
            "required": [
                "fullname",
                "email",
                "password",
                "birthdate"
            ],
            "properties": {
                "fullname": {
                    "type": "string",
                    "uniqueItems": true
                },
                "email": {
                    "type": "string",
                    "uniqueItems": true
                },
                "password": {
                    "type": "string"
                },
                "birthdate": {
                    "type": "string",
                    "format": "date"
                },
                "balance": {
                    "type": "number"
                },
                "points": {
                    "type": "number"
                }
            }
        },
        "CustomerDTO": {
            "required": [
                "id",
                "fullname",
                "email",
                "birthdate"
            ],
            "properties": {
                "id": {
                    "type": "string",
                    "uniqueItems": true
                },
                "fullname": {
                    "type": "string",
                    "uniqueItems": true
                },
                "email": {
                    "type": "string",
                    "uniqueItems": true
                },
                "birthdate": {
                    "type": "string",
                    "format": "date"
                }
            }
        },
        "Customers": {
            "type": "array",
            "$ref": "#/definitions/CustomerDTO"
        },
        "Transaction": {
            "required": [
                "type",
                "amount",
                "receiver"
            ],
            "properties": {
                "type": {
                    "type": "string"
                },
                "amount": {
                    "type": "number"
                },
                "receiver": {
                    "type": "string"
                }
            }
        },
        "Transactions": {
            "type": "array",
            "$ref": "#/definitions/Transaction"
        },
        "Login": {
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        }
    }
}